2010-12-23  Ramon Novoa  <rnovoa@artica.es>
	* build_src.sh, build_windows_agent.sh,
	  build_deb.sh, build_rpm.sh: Return error codes and exclude .po files.



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@3696 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ extras/build_windows_agent.sh
@@ -2,9 +2,13 @@
 CODEHOME=~/code/pandora/trunk
 RPMHOME=/usr/src/packages
 
-mkdir -p $RPMHOME/EXE
+if [ ! -d $RPMHOME/EXE ]; then
+	mkdir -p $RPMHOME/EXE || exit 1
+fi
 
 echo "Creating Pandora FMS Agent Windows installer in $RPMHOME/EXE"
+
+# Windows agent
 rm -rf $CODEHOME/pandora_agents/win32/installer/output/*.exe
 cd $CODEHOME/pandora_agents/win32
 ./build.sh
